% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback_plot}
\alias{feedback_plot}
\title{Feedback plot}
\usage{
feedback_plot(final_scores)
}
\arguments{
\item{final_scores}{(Named list) Contains the scores for the tests.}
}
\description{
Returns a plot displaying the percentage of correct test answers.
}
