% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback_with_graph}
\alias{feedback_with_graph}
\title{Feedback with graph}
\usage{
feedback_with_graph(label, dict)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.}

\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
The participant is given textual and graphical feedback at the end of the test.
}
