% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback_with_score}
\alias{feedback_with_score}
\title{Feedback with score}
\usage{
feedback_with_score(label, dict)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.}

\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textual feedback at the end of the test.
}
