% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{languages}
\alias{languages}
\title{Languages}
\format{
An object of class \code{character} of length 2.
}
\usage{
languages
}
\description{
Lists the languages available for tests.
}
\keyword{datasets}
