% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRT.R
\name{MRT}
\alias{MRT}
\title{MRT}
\usage{
MRT(
  dict,
  item_bank,
  timeout_in_msec = 25000,
  with_practice = TRUE,
  with_feedback = FALSE,
  label = "MRT"
)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{item_bank}{The item_bank of test items used in the MRT.}

\item{timeout_in_msec}{Time in milliseconds until images of a test item disappear.
Defaults to 25000.}

\item{with_practice}{(Logical scalar) Whether to include the training phase.
Defaults to TRUE.}

\item{with_feedback}{(Logical scalar) Whether to display a feedback page.
Defaults to FALSE.}

\item{label}{(Character scalar) Label to give the MRT results in the output file.}
}
\description{
This function defines a MRT module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MRT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
For a standalone implementation of the MRT,
consider using \code{\link{MRT_standalone}()}.
}
