% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{MRT_standalone}
\alias{MRT_standalone}
\title{MRT Standalone
This function launches a standalone testing session for the MRT questionnaire.}
\usage{
MRT_standalone(
  item_bank = OSSAB::MRT_item_bank,
  languages = c("ru", "en"),
  ...
)
}
\arguments{
\item{item_bank}{The item_bank of test items used in the MRT.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{MRT_standalone}()}.}
}
\description{
MRT Standalone
This function launches a standalone testing session for the MRT questionnaire.
}
