% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSSAB.R
\name{OSSAB_battery}
\alias{OSSAB_battery}
\title{OSSAB battery}
\usage{
OSSAB_battery(
  title = "Online Short Spatial Ability Battery",
  tests = c(MRT(), PAT(), PFT(), SRT()),
  languages = OSSAB::languages(),
  dict = OSSAB::OSSAB_dict,
  admin_password = "sirius",
  researcher_email = "tsigeman.es@talantiuspeh.ru or lihanov.mv@talantiuspeh.ru",
  validate_id = "auto",
  ...
)
}
\arguments{
\item{title}{(Character scalar) Title of the test battery.}

\item{tests}{(Character vector)}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"ru"} (Russian), and \code{"en"} (English).
The first language is selected by default.}

\item{dict}{(i18n_dict) The psyquest dictionary used for internationalisation.}

\item{admin_password}{(Character scalar) Password to access the admin panel.}

\item{researcher_email}{(Character scalar) Researcher's email; used in participant help message.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto"
for default validation which means ID should consist only of alphanumeric characters.}

\item{...}{Further arguments to be passed to \code{\link{OSSAB_battery}()}.}
}
\description{
This function defines the OSSAB battery, a series of tests.
Use this function if you want to create a battery of tests.
}
