% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{PAT_standalone}
\alias{PAT_standalone}
\title{PAT Standalone
This function launches a standalone testing session for the PAT questionnaire.}
\usage{
PAT_standalone(
  item_bank = OSSAB::PAT_item_bank,
  languages = c("ru", "en"),
  ...
)
}
\arguments{
\item{item_bank}{The item_bank of test items used in the PAT.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{PAT_standalone}()}.}
}
\description{
PAT Standalone
This function launches a standalone testing session for the PAT questionnaire.
}
