% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{SRT_standalone}
\alias{SRT_standalone}
\title{SRT Standalone
This function launches a standalone testing session for the SRT questionnaire.}
\usage{
SRT_standalone(
  dict = psychSAT::SRT_dict,
  item_bank = psychSAT::SRT_item_bank,
  languages = c("ru", "en"),
  ...
)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{item_bank}{The item_bank of test items used in the SRT.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{SRT_standalone}()}.}
}
\description{
SRT Standalone
This function launches a standalone testing session for the SRT questionnaire.
}
