% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{standalone}
\alias{standalone}
\title{Standalone}
\usage{
standalone(
  label,
  dict = OSSAB::OSSAB_dict,
  item_bank,
  languages = OSSAB::languages,
  admin_password = "sirius",
  researcher_email = "tsigeman.es@talantiuspeh.ru or lihanov.mv@talantiuspeh.ru",
  problems_info = "default",
  validate_id = "auto",
  with_feedback = TRUE,
  with_practice = TRUE,
  ...
)
}
\arguments{
\item{label}{(Scalar character) Three uppercase letter acronym of the questionnaire.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{item_bank}{The item_bank of test items used in the PFT.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"ru"} (Russian), and \code{"en"} (English).
The first language is selected by default.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{problems_info}{Message to display at the bottom of the screen
with advice about what to do if a problem occurs.
The default value, "default", gives
a standard English message including the researcher's email (if provided).
Alternatively, the argument can be either
a) an unnamed character scalar providing a non-internationalised message,
b) a named character vector of internationalised messages with the names
corresponding to language codes,
c) a named list of HTML tag objects providing internationalised messages,
for example:
\code{list(en = shiny::tags$span("Problems? Send an email to ",
                                 shiny::tags$b("researcher@domain.org")),
           de = shiny::tags$span("Probleme? Schreibe eine E-Mail an ",
                                 shiny::tags$b("researcher@domain.org")))}.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto"
for default validation which means ID should consist only of alphanumeric characters.}

\item{with_feedback}{(Scalar boolean) Indicates if performance feedback will be given at the end
of the test. Defaults to TRUE.}

\item{with_practice}{(Boolean scalar) Defines whether instructions and training are included.
Defaults to TRUE.}

\item{...}{Further arguments to be passed to \code{\link{standalone}()}.}
}
\description{
This function launches a standalone testing session for a given questionnaire.
This can be used for data collection, either in the laboratory or online.
}
