% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{title}
\alias{title}
\title{Title}
\format{
An object of class \code{list} of length 2.
}
\usage{
title
}
\description{
Title to be displayed at the top of the page on the login, feedback,
and final page of the battery.
}
\keyword{datasets}
