% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{PFT_standalone}
\alias{PFT_standalone}
\title{PFT Standalone
This function launches a standalone testing session for the PFT questionnaire.}
\usage{
PFT_standalone(
  dict = OSSAB::PFT_dict,
  item_bank = OSSAB::PFT_item_bank,
  languages = c("ru", "en"),
  ...
)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{item_bank}{The item_bank of test items used in the PFT.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), and German (\code{"DE"}).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{PFT_standalone}()}.}
}
\description{
PFT Standalone
This function launches a standalone testing session for the PFT questionnaire.
}
