% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{SRT_standalone}
\alias{SRT_standalone}
\title{SRT Standalone
This function launches a standalone testing session for the SRT questionnaire.}
\usage{
SRT_standalone(
  item_bank = OSSAB::SRT_item_bank,
  languages = OSSAB::languages,
  ...
)
}
\arguments{
\item{item_bank}{The item_bank of test items used in the SRT.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include \code{"ru"} (Russian), and \code{"en"} (English).
The first language is selected by default.}

\item{...}{Further arguments to be passed to \code{\link{SRT_standalone}()}.}
}
\description{
SRT Standalone
This function launches a standalone testing session for the SRT questionnaire.
}
